+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Removes metadata cache provider [#654](https://github.com/parres-hq/whitenoise_flutter/pull/654)
+
 ### Fixed
 
 - Fixed 2 users group creation when DM already exists
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved haptic feedback for chat context menus.
 - Pinned the header in auth flow screens when keyboard is open for easier back navigation.
 - Fixed broken profile image upload [#701](https://github.com/parres-hq/whitenoise_flutter/pull/701)
+- Fixed double rendering issue for some messages [#654](https://github.com/parres-hq/whitenoise_flutter/pull/654)
 
 ### Security
 
@@ -40,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update group name and description from group details screen
 - Add back buttons in auth flow screens for easier navigation
 - Paste from clipboard functionality in new chat bottom sheet.
+- Pin chats to the top on the chat list screen
 
 ### Changed